refactor: decompose OrbViewPage and CvExportPage god components (#181)
Both files were flagged in #181.7 as god components handling too many
concerns. OrbViewPage was 2111 lines, CvExportPage was 1670 lines.

Extractions:

1. OrbViewPage → SharePanel.tsx (1090 lines)
   The share/grant/visibility management panel was an inline function
   component spanning lines 67-1136. Moved to its own file at
   components/graph/SharePanel.tsx with its own imports. OrbViewPage
   drops from 2111 to 1027 lines (-51%).

2. CvExportPage → cv-export-utils.ts (139 lines)
   Pure utility functions and constants (parseDateSort, sortDesc,
   collectBlocks, computePageBreaks, hexToHsl, section types/labels/
   order) extracted to a co-located utils module. These are trivially
   unit-testable now. CvExportPage drops from 1670 to 1542 (-8%).

CvExportPage's renderSection (10 section renderers, ~250 lines) and
the drag-drop logic are tightly coupled to parent state and are left
in place for now — extracting them would require prop-drilling or a
context provider that adds complexity without reducing it.

No behavioral changes. Frontend lint: 0 errors, 35 warnings (all
pre-existing).
